Working signup smoketest hourly at 20 past

+++ smoketest/delete-smoketest-apps.rb
@@ -1,10 +1,19 @@
-require('Oktakit');
+require 'rubygems'
+require 'bundler/setup'
+require 'oktakit'
 
-id = ARGV[0];
+id = ARGV[0]
 
-client = Oktakit.new(token: ENV['OKTA_TOKEN'], api_endpoint: ENV['OKTA_BASE_URL']);
-response, http_status = client.deactivate_application(id);
-p(http_status);
-response, http_status = client.delete_application(id);
-p(http_status);
-p(response);
+client = Oktakit.new(token: ENV['OKTA_TOKEN'], api_endpoint: ENV['OKTA_BASE_URL'])
+response, http_status = client.deactivate_application(id)
+p http_status
+if http_status != 200
+  p "The app was not able to be deactivated. This is a problem."
+  exit 1
+end
+response, http_status = client.delete_application(id)
+p http_status
+if http_status != 204
+  p "The app was not able to be deleted. This is a problem."
+  exit 1
+end
